# src/swagger/swagger-latest.yaml
swagger: '2.0'
info:
  contact: {}
paths:
...
  /api/v2/lists/{list_id}/posters:
    get:
      description: get posters in list
      tags:
        - lists/v2
      summary: Get posters in list
      parameters:
        - type: string
          description: JWT-format token
          name: X-User-Token
          in: header
          required: true
        - type: integer
          description: ListId
          name: list_id
          in: path
          required: true
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ListPosterResponse'
        '400':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '404':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
  /api/v2/lists/{list_id}/posters/{poster_id}:
    put:
      description: change poster position in list
      consumes:
        - application/json
      tags:
        - lists/v2
      summary: Change poster position in list
      parameters:
        - type: string
          description: JWT-format token
          name: X-User-Token
          in: header
          required: true
        - type: integer
          description: ListId
          name: list_id
          in: path
          required: true
        - type: integer
          description: PosterId
          name: poster_id
          in: path
          required: true
        - description: Change position body
          name: position
          in: body
          required: true
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ListPositionRequest'
      responses:
        '200':
          description: OK
        '400':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '404':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
    post:
      description: Adds poster in list. If poster already exists in some list, it will be moved to new list.
      tags:
        - lists/v2
      summary: Add poster in list
      parameters:
        - type: string
          description: JWT-format token
          name: X-User-Token
          in: header
          required: true
        - type: integer
          description: ListId
          name: list_id
          in: path
          required: true
        - type: integer
          description: PosterId
          name: poster_id
          in: path
          required: true
      responses:
        '200':
          description: Poster moved
        '201':
          description: Poster added
        '400':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '404':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
    delete:
      description: delete poster from list
      consumes:
        - application/json
      tags:
        - lists/v2
      summary: Delete poster from list
      parameters:
        - type: string
          description: JWT-format token
          name: X-User-Token
          in: header
          required: true
        - type: integer
          description: ListId
          name: list_id
          in: path
          required: true
        - type: integer
          description: PosterId
          name: poster_id
          in: path
          required: true
      responses:
        '200':
          description: OK
        '400':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '401':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '404':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        '500':
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
...
definitions:
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse:
    type: object
    properties:
      error:
        type: string
        example: error description
...
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ListPositionRequest:
    type: object
    properties:
      position:
        type: integer
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ListPosterResponse:
    type: object
    properties:
      listId:
        description: ID       int `json:"id"`
        type: integer
      position:
        type: integer
      posterId:
        type: integer
...